feat: logique permettant de check la date limite d'utilisation du lien d'invitation

+++ backend/migrations/20260105172255_add_banquet_time_to_events.js
@@ -1,11 +1,19 @@
-exports.up = function (knex) {
-  return knex.schema.alterTable('EVENTS', function (table) {
-    table.time('banquet_time').nullable();
-  });
+exports.up = async function (knex) {
+  //Évite l’erreur Duplicate column name
+  //Migration idempotente
+  const hasColumn = await knex.schema.hasColumn('EVENTS', 'banquet_time');
+  if(!hasColumn){
+    return knex.schema.table('EVENTS', function (table) {
+      table.time('banquet_time').nullable();
+    });
+  }
 };
 
-exports.down = function (knex) {
-  return knex.schema.alterTable('EVENTS', function (table) {
-    table.dropColumn('banquet_time');
-  });
+exports.down = async function (knex) {
+  const hasColumn = await knex.schema.hasColumn('EVENTS', 'banquet_time');
+  if(!hasColumn){
+    return knex.schema.table('EVENTS', function (table) {
+      table.dropColumn('banquet_time');
+    });
+  }
 };
